chore: separate database function files
